Git - "Close Repository" and "Close Other Repositories" now works in the main Source Control view (#202521)
